+++ NEW_Z_BOX/new_dsp_z/device/driverlib/gpio.h
@@ -5,10 +5,8 @@
 // TITLE:  C28x GPIO driver.
 //
 //###########################################################################
-// 
-// C2000Ware v6.00.00.00
-//
-// Copyright (C) 2024 Texas Instruments Incorporated - http://www.ti.com
+// $Copyright:
+// Copyright (C) 2022 Texas Instruments Incorporated - http://www.ti.com
 //
 // Redistribution and use in source and binary forms, with or without 
 // modification, are permitted provided that the following conditions 
@@ -305,8 +303,8 @@ GPIO_getInterruptType(GPIO_ExternalIntNum extIntNum)
     //
     // Read the selected polarity from the appropriate register.
     //
-    return((GPIO_IntType)((uint16_t)(HWREGH(XINT_BASE + (uint16_t)extIntNum) &
-                                     XINT_1CR_POLARITY_M)));
+    return((GPIO_IntType)(HWREGH(XINT_BASE + (uint16_t)extIntNum) &
+                          XINT_1CR_POLARITY_M));
 }
 
 //*****************************************************************************
