+++ NEW_Z_BOX/new_dsp_z/device/driverlib/sci.h
@@ -5,10 +5,8 @@
 // TITLE:  C28x SCI driver.
 //
 //###########################################################################
-// 
-// C2000Ware v6.00.00.00
-//
-// Copyright (C) 2024 Texas Instruments Incorporated - http://www.ti.com
+// $Copyright:
+// Copyright (C) 2022 Texas Instruments Incorporated - http://www.ti.com
 //
 // Redistribution and use in source and binary forms, with or without 
 // modification, are permitted provided that the following conditions 
@@ -386,10 +384,10 @@ SCI_lockAutobaud(uint32_t base)
 //! \param txLevel is the transmit FIFO interrupt level, specified as one of
 //! the following:
 //! \b SCI_FIFO_TX0, \b SCI_FIFO_TX1, \b SCI_FIFO_TX2, . . . or
-//! \b SCI_FIFO_TX16.
+//! \b SCI_FIFO_TX15.
 //! \param rxLevel is the receive FIFO interrupt level, specified as one of
 //! the following
-//! \b SCI_FIFO_RX0, \b SCI_FIFO_RX1, \b SCI_FIFO_RX2, ... or \b SCI_FIFO_RX16.
+//! \b SCI_FIFO_RX0, \b SCI_FIFO_RX1, \b SCI_FIFO_RX2, ... or \b SCI_FIFO_RX15.
 //!
 //! This function sets the FIFO level at which transmit and receive interrupts
 //! are generated.
@@ -422,10 +420,10 @@ SCI_setFIFOInterruptLevel(uint32_t base, SCI_TxFIFOLevel txLevel,
 //! \param base is the base address of the SCI port.
 //! \param txLevel is a pointer to storage for the transmit FIFO interrupt
 //! level, returned as one of the following:
-//! \b SCI_FIFO_TX0, \b SCI_FIFO_TX1, \b SCI_FIFO_TX2, ... or \b SCI_FIFO_TX16.
+//! \b SCI_FIFO_TX0, \b SCI_FIFO_TX1, \b SCI_FIFO_TX2, ... or \b SCI_FIFO_TX15.
 //! \param rxLevel is a pointer to storage for the receive FIFO interrupt
 //! level, returned as one of the following:
-//! \b SCI_FIFO_RX0, \b SCI_FIFO_RX1, \b SCI_FIFO_RX2, ... or \b SCI_FIFO_RX16.
+//! \b SCI_FIFO_RX0, \b SCI_FIFO_RX1, \b SCI_FIFO_RX2, ... or \b SCI_FIFO_RX15.
 //!
 //! This function gets the FIFO level at which transmit and receive interrupts
 //! are generated.
@@ -490,15 +488,15 @@ SCI_getConfig(uint32_t base, uint32_t lspclkHz, uint32_t *baud,
     // Compute the baud rate.
     //
     *baud = lspclkHz /
-            ((1U + (((uint32_t)HWREGH(base + SCI_O_HBAUD) << 8U) |
+            ((1U + (uint32_t)((uint32_t)(HWREGH(base + SCI_O_HBAUD) << 8U) |
                HWREGH(base + SCI_O_LBAUD))) * 8U);
 
     //
     // Get the parity, data length, and number of stop bits.
     //
-    *config = (uint32_t)HWREGH(base + SCI_O_CCR) & (SCI_CONFIG_PAR_MASK |
-                                                    SCI_CONFIG_STOP_MASK |
-                                                    SCI_CONFIG_WLEN_MASK);
+    *config = HWREGH(base + SCI_O_CCR) & (SCI_CONFIG_PAR_MASK |
+                                          SCI_CONFIG_STOP_MASK |
+                                          SCI_CONFIG_WLEN_MASK);
 }
 
 //*****************************************************************************
@@ -1026,7 +1024,7 @@ SCI_isTransmitterBusy(uint32_t base)
         // With FIFO enhancement, determine if the SCI is busy.
         //
         return(((HWREGH(base + SCI_O_FFTX) & SCI_FFTX_TXFFST_M) !=
-                 0U) ? true : false);
+                 0) ? true : false);
     }
     else
     {
@@ -1127,6 +1125,10 @@ SCI_writeCharBlockingNonFIFO(uint32_t base, uint16_t data)
 //! to determine if the transmit buffer or FIFO have space available.
 //! \e data is a uint16_t but only 8 bits are written to the SCI port.  SCI
 //! only transmits 8 bit characters.
+//!
+//! This function replaces the original SCICharNonBlockingPut() API and
+//! performs the same actions.  A macro is provided in <tt>sci.h</tt> to map
+//! the original API to this API.
 //!
 //! \return None.
 //
@@ -1267,6 +1269,10 @@ SCI_readCharBlockingNonFIFO(uint32_t base)
 //! function does not block and only reads the receive buffer.  The user should
 //! use SCI_isDataAvailableNonFIFO() or SCI_getRxFIFOStatus() to determine if
 //! the receive buffer or FIFO have data available.
+//!
+//! This function replaces the original SCICharNonBlockingGet() API and
+//! performs the same actions.  A macro is provided in <tt>sci.h</tt> to map
+//! the original API to this API.
 //!
 //! \return Returns \e uin16_t which is read from the receive buffer.
 //
